minor refactoring/renaming
