+++ api/src/main/java/org/apache/iceberg/expressions/ResidualEvaluator.java
@@ -28,6 +28,7 @@ import org.apache.iceberg.PartitionSpec;
 import org.apache.iceberg.StructLike;
 import org.apache.iceberg.expressions.ExpressionVisitors.BoundExpressionVisitor;
 import org.apache.iceberg.transforms.Transform;
+import org.apache.iceberg.util.NaNUtil;
 
 /**
  * Finds the residuals for an {@link Expression} the partitions in the given {@link PartitionSpec}.
@@ -154,12 +155,12 @@ public class ResidualEvaluator implements Serializable {
 
     @Override
     public <T> Expression isNaN(BoundReference<T> ref) {
-      return NaNUtils.isNaN(ref.eval(struct)) ? alwaysTrue() : alwaysFalse();
+      return NaNUtil.isNaN(ref.eval(struct)) ? alwaysTrue() : alwaysFalse();
     }
 
     @Override
     public <T> Expression notNaN(BoundReference<T> ref) {
-      return NaNUtils.isNaN(ref.eval(struct)) ? alwaysFalse() : alwaysTrue();
+      return NaNUtil.isNaN(ref.eval(struct)) ? alwaysFalse() : alwaysTrue();
     }
 
     @Override
